+++ packages/react/src/index.ts
@@ -1,5 +1,6 @@
 export * from './hooks/juice-contracts/read';
 export * from './hooks/juice-contracts/write';
 export * from './hooks/juice-contracts/contracts';
+export * from './types';
 
 export { JuiceProvider } from './contexts/JuiceContext';
