+++ packages/react/src/types/contractReadHookResponse.ts
@@ -1,0 +1,5 @@
+export interface ContractReadHookResponse<T> {
+  data: T | undefined;
+  loading: boolean;
+  error: Error | undefined;
+}
