+++ packages/react/src/types/projectId.ts
@@ -1,0 +1,1 @@
+export type ProjectId = number;
